Flash-Lang-main-RevA/fastapi-backend/requirements.txt
fastapi
uvicorn[standard]
requests
youtube-transcript-api
beautifulsoup4
python-dotenv
pydantic
